+++ src/main/java/racinggame/domain/RacingCars.java
@@ -3,6 +3,7 @@ package racinggame.domain;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import racinggame.domain.validator.Validator;
 
 public class RacingCars {
